modified step def names and removed extra spaces from classes
