+++ src/test/java/com/chroma/features/AddHomeworkModule.feature
@@ -1,9 +1,8 @@
-@tag
 Feature: Add Homework Module
 
 @homework
 Scenario: Inspection Of Add Homework Module Link
 Given a user is on the login page "https://mexil.it/chroma_tech_academy/dev/site/login"
 When user logs in with valid credentials username "general@teacher.com" and password "123456"
-And user clicks the homework dropdown on dashboard page "https://mexil.it/chroma_tech_academy/dev/admin/admin/dashboard"
+And user clicks the student information dropdown on dashboard page "https://mexil.it/chroma_tech_academy/dev/admin/admin/dashboard"
 Then The following modules should be displayed "Add Homework"
