+++ src/test/java/com/chroma/runners/Runners.java
@@ -1,7 +1,6 @@
 package com.chroma.runners;
 
 import org.junit.runner.RunWith;
-
 import cucumber.api.CucumberOptions;
 import cucumber.api.junit.Cucumber;
 
@@ -11,7 +10,7 @@ import cucumber.api.junit.Cucumber;
 		"pretty", "com.aventstack.extentreports.cucumber.adapter.ExtentCucumberAdapter:"  }, 
 		features = "src/test/java/com/chroma/features", 
 		glue = "com.chroma.stepDefinitions", 
-		tags = "@homework",
+		tags = "@homework", 
 		dryRun =  false, 
 		monochrome = true, 
 		strict = true
